Refactor private methods to protected.
This is driven by my belief a library must allow for consumers to be
open to extending the methods to suit their own needs. In the case of
authentication there are an endless list of potential scenarios where
users may need to handle scopes differently then the root class defines.
This is only one example.

Cleaned up namespace statements and use declarations to be PSR1
compatable.